requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
dnspython==2.2.1
idna==3.4
lxml==4.9.2
nslookup==1.6.1
pygeoip==0.3.2
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.13
whois==0.9.21
